# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MindTask API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List workspaces for current user
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceResponse"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace by ID
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [workspaces]
      summary: List workspace members
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMemberResponse"
    post:
      operationId: addWorkspaceMember
      tags: [workspaces]
      summary: Add member to workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMemberResponse"
  /workspaces/{workspaceId}/members/{memberId}:
    put:
      operationId: updateWorkspaceMember
      tags: [workspaces]
      summary: Update member role
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRoleRequest"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMemberResponse"
    delete:
      operationId: removeWorkspaceMember
      tags: [workspaces]
      summary: Remove member from workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/maps:
    get:
      operationId: listMaps
      tags: [maps]
      summary: List maps in workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of maps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MapResponse"
    post:
      operationId: createMap
      tags: [maps]
      summary: Create a map
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMapRequest"
      responses:
        "201":
          description: Map created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapResponse"
  /workspaces/{workspaceId}/maps/{mapId}:
    get:
      operationId: getMap
      tags: [maps]
      summary: Get map with cards and connections
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Map with cards and connections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapDetailResponse"
    put:
      operationId: updateMap
      tags: [maps]
      summary: Update map name
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMapRequest"
      responses:
        "200":
          description: Updated map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapResponse"
    delete:
      operationId: deleteMap
      tags: [maps]
      summary: Delete map
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mapId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /workspaces/{workspaceId}/dashboard:
    get:
      operationId: getDashboard
      tags: [workspaces]
      summary: Get dashboard summary for a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    CardVisualStatus:
      type: string
      enum: [no_task, pending, in_progress, completed, overdue]
    WorkspaceRole:
      type: string
      enum: [admin, editor, executor]
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, createdAt]
...
    CreateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
    WorkspaceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        role:
          $ref: "#/components/schemas/WorkspaceRole"
      required: [id, name, createdBy, createdAt, role]
    WorkspaceDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        role:
          $ref: "#/components/schemas/WorkspaceRole"
        members:
          type: array
          items:
            $ref: "#/components/schemas/WorkspaceMemberResponse"
        maps:
          type: array
          items:
            $ref: "#/components/schemas/MapResponse"
      required: [id, name, createdBy, createdAt, role, members, maps]
    WorkspaceMemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/WorkspaceRole"
        user:
          $ref: "#/components/schemas/UserResponse"
        createdAt:
          type: string
          format: date-time
      required: [id, workspaceId, userId, role, user, createdAt]
    AddMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/WorkspaceRole"
      required: [email, role]
    UpdateMemberRoleRequest:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/WorkspaceRole"
      required: [role]
    CreateMapRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]
    MapResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, workspaceId, name, createdBy, createdAt, updatedAt]
    MapDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        name:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        cards:
          type: array
          items:
            $ref: "#/components/schemas/CardResponse"
        connections:
          type: array
          items:
            $ref: "#/components/schemas/ConnectionResponse"
      required: [id, workspaceId, name, createdBy, createdAt, updatedAt, cards, connections]
...
    CardResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mapId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        positionX:
          type: number
        positionY:
          type: number
        statusVisual:
          $ref: "#/components/schemas/CardVisualStatus"
        taskId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, mapId, title, positionX, positionY, statusVisual, createdAt, updatedAt]
...
    ConnectionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mapId:
          type: string
          format: uuid
        sourceCardId:
          type: string
          format: uuid
        targetCardId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
      required: [id, mapId, sourceCardId, targetCardId, createdAt]
...
    DashboardResponse:
      type: object
      properties:
        workspaceId:
          type: string
          format: uuid
        totalMaps:
          type: integer
        totalCards:
          type: integer
        totalTasks:
          type: integer
        tasksByStatus:
          type: object
          properties:
            pending:
              type: integer
            in_progress:
              type: integer
            completed:
              type: integer
            overdue:
              type: integer
          required: [pending, in_progress, completed, overdue]
        tasksByPriority:
          type: object
          properties:
            low:
              type: integer
            medium:
              type: integer
            high:
              type: integer
            critical:
              type: integer
          required: [low, medium, high, critical]
      required: [workspaceId, totalMaps, totalCards, totalTasks, tasksByStatus, tasksByPriority]