# lib/api-zod/src/generated/types/mapResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindTask API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MapResponse {
  id: string;
  workspaceId: string;
  name: string;
  createdBy: string;
  createdAt: Date;
  updatedAt: Date;
}