# lib/api-zod/src/generated/types/updateMemberRoleRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindTask API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WorkspaceRole } from "./workspaceRole";

export interface UpdateMemberRoleRequest {
  role: WorkspaceRole;
}