# lib/api-zod/src/generated/types/workspaceMemberResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindTask API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserResponse } from "./userResponse";
import type { WorkspaceRole } from "./workspaceRole";

export interface WorkspaceMemberResponse {
  id: string;
  workspaceId: string;
  userId: string;
  role: WorkspaceRole;
  user: UserResponse;
  createdAt: Date;
}